requirements.txt
six==1.11.0
numpy==1.14.1
tensorflow==1.6.0
Django==1.11.10
Image==1.5.19
Pillow==5.0.0
djangorestframework==3.7.7
requests
